JPEGFileImageProvider and a bunch of other things

+++ src/Texture.hpp
@@ -9,20 +9,20 @@
 struct Image;
 
 struct Tile {
-    unsigned int id;
+    unsigned id;
     int x, y;
-    int w, h;
-    unsigned int format;
+    size_t w, h;
+    unsigned format;
 };
 
 struct Texture {
     std::vector<Tile> tiles;
     ImVec2 size;
-    unsigned int format = -1;
+    unsigned format = -1;
 
     ~Texture();
 
-    void create(int w, int h, unsigned int format);
+    void create(size_t w, size_t h, unsigned format);
     void upload(const Image* img, ImRect area);
     ImVec2 getSize() { return size; }
 };
